Begin breaking out field types
